requirements.txt
absl-py==1.4.0
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.37.3
botocore==1.37.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.18.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.52.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.5.2
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multidict==6.4.3
mysql-connector-python==8.2.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==8.13.49
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==4.21.12
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyOpenSSL==25.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytimeparse==1.1.8
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
s3fs==2025.3.2
s3transfer==0.11.3
scikit-learn==1.5.2
scipy==1.15.3
selenium==4.23.1
Send2Trash==1.8.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-ml-python==1.8.3
snowflake-snowpark-python==1.31.1
snowflake.core==1.4.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
Wikipedia-API==0.8.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.4
yarg==0.1.9
yarl==1.20.0
